# apps/docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mailmon API",
    "description": "Development documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.mailmon.dev",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "NonEmptyString": {
        "type": "string",
        "description": "a non empty string",
        "title": "nonEmptyString",
        "minLength": 1
      }
    }
  },
  "paths": {
    "/v1/mailboxes/connect-sessions": {
      "post": {
        "operationId": "postV1MailboxesConnectSessions",
        "summary": "Create a mailbox connect session",
        "responses": {
          "201": {
            "description": "Connect session created"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["provider", "tenantExternalId", "mailboxExternalId", "redirectUrl"],
                "properties": {
                  "provider": {
                    "type": "string",
                    "enum": ["gmail"]
                  },
                  "tenantExternalId": {
                    "$ref": "#/components/schemas/NonEmptyString"
                  },
                  "mailboxExternalId": {
                    "$ref": "#/components/schemas/NonEmptyString"
                  },
                  "redirectUrl": {
                    "$ref": "#/components/schemas/NonEmptyString"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/v1/webhook-endpoints": {
      "post": {
        "operationId": "postV1WebhookEndpoints",
        "summary": "Create a webhook endpoint",
        "responses": {
          "201": {
            "description": "Webhook endpoint created"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": {
                    "$ref": "#/components/schemas/NonEmptyString"
                  },
                  "description": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/NonEmptyString"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/v1/webhook-endpoints/{endpointId}/subscriptions": {
      "post": {
        "operationId": "postV1WebhookEndpointsByEndpointIdSubscriptions",
        "summary": "Create mailbox-scoped webhook subscriptions",
        "responses": {
          "201": {
            "description": "Webhook subscriptions created"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Mailbox or webhook endpoint not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mailboxIds", "eventTypes"],
                "properties": {
                  "mailboxIds": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "$ref": "#/components/schemas/NonEmptyString"
                    }
                  },
                  "eventTypes": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                      "type": "string",
                      "enum": ["message.created", "message.updated", "thread.updated"]
                    }
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "endpointId",
            "required": true
          }
        ]
      }
    },
    "/v1/mailboxes/{mailboxId}/sync-runs": {
      "get": {
        "operationId": "getV1MailboxesByMailboxIdSyncRuns",
        "summary": "List mailbox sync runs",
        "responses": {
          "200": {
            "description": "Mailbox sync runs"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            },
            "description": "a string to be decoded into a number"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "mailboxId",
            "required": true
          }
        ]
      }
    },
    "/v1/replays": {
      "post": {
        "operationId": "postV1Replays",
        "summary": "Create a mailbox event replay",
        "responses": {
          "201": {
            "description": "Replay created"
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Overlapping active replay conflict"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["mailboxId", "webhookEndpointId", "startTime", "endTime"],
                "properties": {
                  "mailboxId": {
                    "$ref": "#/components/schemas/NonEmptyString"
                  },
                  "webhookEndpointId": {
                    "$ref": "#/components/schemas/NonEmptyString"
                  },
                  "startTime": {
                    "$ref": "#/components/schemas/NonEmptyString"
                  },
                  "endTime": {
                    "$ref": "#/components/schemas/NonEmptyString"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/v1/messages": {
      "get": {
        "operationId": "getV1Messages",
        "summary": "List mailbox messages",
        "responses": {
          "200": {
            "description": "Mailbox messages"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            },
            "description": "a string to be decoded into a number"
          },
          {
            "in": "query",
            "name": "mailboxId",
            "schema": {
              "$ref": "#/components/schemas/NonEmptyString"
            }
          }
        ]
      }
    },
    "/v1/threads": {
      "get": {
        "operationId": "getV1Threads",
        "summary": "List mailbox threads",
        "responses": {
          "200": {
            "description": "Mailbox threads"
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "string"
            },
            "description": "a string to be decoded into a number"
          },
          {
            "in": "query",
            "name": "mailboxId",
            "schema": {
              "$ref": "#/components/schemas/NonEmptyString"
            }
          }
        ]
      }
    }
  }
}